+++ packages/popup/package.json
@@ -44,8 +44,7 @@
     ]
   },
   "dependencies": {
-    "@floating-ui/react": "^0.25.4",
-    "@titicaca/core-elements": "workspace:*"
+    "@titicaca/tds-ui": "workspace:*"
   },
   "devDependencies": {
     "react": "^18.2.0",
