+++ packages/tds-ui/src/components/popup/index.ts
@@ -1,0 +1,1 @@
+export * from './popup'
